+++ Data/Retail/ExportedWidgetEvents.lua
@@ -24,7 +24,7 @@ local a = addon; -- Saves some characters and file size as we use this a lot in 
 -- [[ Everything after these lines is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. - AUTOGENTOKEN ]] --
 
--- [[ Exported at 2023-11-09 21-46-29 ]] --
+-- [[ Exported at 2023-11-11 01-05-01 ]] --
 tasks = {
     {N, 4729, 1393, 4240492, a.L["Community Feast: Next"]}, -- Community Feast: Next
     {N, 4731, 1393, 4240492, a.L["Community Feast: Active"]}, -- Community Feast: Active
@@ -35,13 +35,13 @@ tasks = {
 };
 
 function exportedWidgetEvents.InjectDynamicOptions()
-    KrowiAF_RegisterEventOptions("Widget", "Community_Feast", a.L["Community Feast"], {4729}, a.L["Community Feast: Next"]);
-    KrowiAF_RegisterEventOptions("Widget", "Community_Feast", a.L["Community Feast"], {4731}, a.L["Community Feast: Active"]);
-    KrowiAF_RegisterDeSelectAllEventOptions("Widget", "Community_Feast", { 4729, 4731 });
-    KrowiAF_RegisterEventOptions("Widget", "Time_Rifts", a.L["Time Rifts"], {4987}, a.L["Time Rift: Active"]);
-    KrowiAF_RegisterEventOptions("Widget", "Time_Rifts", a.L["Time Rifts"], {4992}, a.L["Time Rift: Next"]);
-    KrowiAF_RegisterDeSelectAllEventOptions("Widget", "Time_Rifts", { 4987, 4992 });
-    KrowiAF_RegisterEventOptions("Widget", "Superbloom", a.L["Superbloom"], {5157}, a.L["Superbloom: Next"]);
-    KrowiAF_RegisterEventOptions("Widget", "Superbloom", a.L["Superbloom"], {5323}, a.L["Superbloom: Active"]);
-    KrowiAF_RegisterDeSelectAllEventOptions("Widget", "Superbloom", { 5157, 5323 });
+    KrowiAF_RegisterEventOptions("Widget", {4729}, a.L["Community Feast: Next"], a.L["Community Feast"], 5, a.L["Dragonflight"]);
+    KrowiAF_RegisterEventOptions("Widget", {4731}, a.L["Community Feast: Active"], a.L["Community Feast"], 5, a.L["Dragonflight"]);
+    KrowiAF_RegisterDeSelectAllEventOptions("Widget", {4729, 4731}, a.L["Community Feast"], "Dragonflight");
+    KrowiAF_RegisterEventOptions("Widget", {4987}, a.L["Time Rift: Active"], a.L["Time Rifts"], 13, a.L["Dragonflight"]);
+    KrowiAF_RegisterEventOptions("Widget", {4992}, a.L["Time Rift: Next"], a.L["Time Rifts"], 13, a.L["Dragonflight"]);
+    KrowiAF_RegisterDeSelectAllEventOptions("Widget", {4987, 4992}, a.L["Time Rifts"], "Dragonflight");
+    KrowiAF_RegisterEventOptions("Widget", {5157}, a.L["Superbloom: Next"], a.L["Superbloom"], 15, a.L["Dragonflight"]);
+    KrowiAF_RegisterEventOptions("Widget", {5323}, a.L["Superbloom: Active"], a.L["Superbloom"], 15, a.L["Dragonflight"]);
+    KrowiAF_RegisterDeSelectAllEventOptions("Widget", {5157, 5323}, a.L["Superbloom"], "Dragonflight");
 end