+++ CHANGELOG.md
@@ -1,18 +1,21 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## 65.1 - 2023-11-07
+## 66.0 - 2023-11-12
 ### Added
 - Outland Cup event
-- Adjustable width for the achievements
+- Adjustable width for the achievements and summary
 
 ### Changed
 - Both Researchers Under Fire events are now showing the 'Next' and 'Active' timers instead of the entire week when in Zaralek Cavern (outside will show that no time data is available)
 - Dreamsurges are now showing the 'Next' timer for the next portal instead of the entire week
+- Widget events are now also grouped under the world events as this makes more sense for the end user
+- Events are now grouped per expansion instead of having a prefix
 
 ### Fixed
 - Dreamsurges are now properly tracked (new Ids so settings might need to be reapplyed)
 - Fyrakk Assaults should once again properly track
+- Potential fix for Excluded achievements resetting
 
 ## 65.0 - 2023-11-07
 ### Added